Always use -fmacro-prefix-map
When `deterministic_paths` is set, we are currently using
`-ffile-prefix-map` to produce the same path in data and debug info. In
the case of absolute paths, their emscripten path is replaced with a
fake path `/emsdk/emscripten`, and in the case of relative paths, all
path relative to the emscripten directory is removed, so
`../../system/lib/somefile.c` becomes `system/lib/somefiles.c`.
https://github.com/emscripten-core/emscripten/blob/f66b5d706e174d9e5cc6122c06ea29dcd2735cd0/tools/system_libs.py#L472-L477
https://github.com/emscripten-core/emscripten/blob/f66b5d706e174d9e5cc6122c06ea29dcd2735cd0/tools/system_libs.py#L495-L501

But this does not make relative paths and absolute paths the same, which
can be a problem when data generated by `__FILE__` macro is included in
one of code size tests. This problem is discussed in #23195.

This PR makes `__FILE__` macro produce the same data in all cases by
using the fake path `/emsdk/emscripten` as its base, so that it wouldn't
change any results for code size tests. This is done by
`-fmacro-prefix-map`. This differs from the current behavior because we
don't handle relative and absolute paths differently.

For the debug info, when `deterministic_paths` is set, this uses a fake
path `/emsdk/emscripten` as a base emscripten path. When
`deterministic_paths` is not set, this uses real local absolute paths in
the debug info. This allows local developers to see their real paths in
the debug info while continuing to use the same (fake) path
`/emsdk/emscripten` we have used so far for the release binaries. Users
can set their debug base path to whatever path they like, but given that
we have used `/emsdk/emscripten` in release binaries for a while, it is
possible that some users have set their configuration with this
directory, so it would be better not to break them by changing it. This
is done by `-ffile-prefix-map` as we have done so far, which is an alias
for both `-fdebug-prefix-map` and `-fmacro-prefix-map`.

This is basically implementing what's suggested in
https://github.com/emscripten-core/emscripten/issues/23195#issuecomment-2549784273
and
https://github.com/emscripten-core/emscripten/issues/23195#issuecomment-2549859260

This also turns `deterministic_paths` on for the Ninja path in
embuilder for consistency with the non-Ninja path.

Fixes #23915.