Remove unnecessary @nocollape. NFC
This was added back in #20191 but I can't see how it can work.  The
thing is being applied to here is a function variable. e.g.:

```
var Modulle = function() { ... return the a new module };
```

From the documentation for @nocollape is not clear to me that it has
any effect on normal variables like this.