Convert `createWasm` async/await where possible (#23157)
The advantage if using `await` in the cases where we can is that it
avoids the generation or wrapper function for each export.

So instead of:

```
var wasmExport = createWasm(); // returns empty object
...
var malloc = (..) => (malloc = wasmExports['malloc'])(..);
```

We can generate:

```
var wasmExport = await createWasm(); // returns actual exports
...
var malloc = wasmExports['malloc'];
```

This only currently works in MODULARIZE mode where the code is running
inside a factory function. Otherwise it would end up using
top-level-await.

One wrinkle here is that this is not currently supported when closure is
enabled because we run closure only on the contents of the factory
function so closure ends up seeing this as a top level await when its
not.

There are two minor observable effects of this change:

1. In `MODULARIZE` mode its no longer possible to call `new Foo()` with
factory function. We already added a debug error for in #23210.
2. Because we now can `await` for createWasm to return, the `run` method
can run on first call, which means it runs `main` on first call, which
means main will now run before `postjs` code, just like it would with
sync instantiation.