Remove test_export_es6_allows_export_in_post_js. NFC
This test was added in #13937.  Here we are testing something that
we actually decided we did not want to support.  Specifically in #13894
we decided that `--extern-post-js` was the place to allow this.

More importantly the resulting JS module simple doesn't work because in
`MODULARIZE` mode the code from `--post-js` does not end up at the top
level but instead inside a factory function.  Note that this test does
not try to exectute the resulting JS.  If you try to execute it you get
`SyntaxError: Unexpected token 'export'`.

Since we don't support it I think its best to remove this test.

I'm also working on a change to move the modularization code before
acorn and closure, and with that change this won't even compile (because
the modularization process puts all the code inside and inner scope.