[contrib.glfw3] added support for pthread (#23245)
Following some user request, I made the necessary changes to the code as
well as the port file to add pthread support. I followed the pattern in
other ports (like sdl2) for adding the `-pthread` option when
`settings.PTHREAD` is defined and used the same extension for the
library name (`-mt`).