Always enable deterministic_paths
This removes `deterministic_paths` option by turning it on all the time,
in order to produce reproducible builds, both for `__FILE__` macro and
debug info paths.

This is an alternative to #23212, which did not remove
`deterministic_paths` but always set only `-fmacro-prefix-map` on.

This PR is what was suggested in
https://github.com/emscripten-core/emscripten/issues/23195#issuecomment-2549861247
and
https://github.com/emscripten-core/emscripten/pull/23212#discussion_r1891058881.

Fixes #23195.