Remove dummy emscripten_set_offscreencanvas_size. NFC
This is an internal-only API that is not directly callable by user code.

The only place that this function is called is from
`system/lib/html5/offscreencanvas.c` as a result of
`_emscripten_set_offscreencanvas_size_on_thread`.

The only call to `_emscripten_set_offscreencanvas_size_on_thread`
is from `setOffscreenCanvasSizeOnTargetThread` which is only ever
included when `OFFSCREENCANVAS_SUPPORT` is enabled.