Update libcxx and libcxxabi to LLVM 19.1.4 (#22994)
This updates libcxx and libcxxabi to LLVM 19.1.4:
https://github.com/llvm/llvm-project/releases/tag/llvmorg-19.1.4

The initial update was done using `update_libcxx.py` and
`update_libcxxabi.py`, and subsequent fixes were made in indidual
commits. The commit history here is kind of messy because of CI testing
so not all individual commits are noteworthy.

Additional changes:

- Build libcxx and libcxxabi with C++23:
https://github.com/emscripten-core/emscripten/pull/22994/commits/8b0bfdf2d3039a3d6799dd7b407595f3ad246623

https://github.com/llvm/llvm-project/blob/aadaa00de76ed0c4987b97450dd638f63a385bed/libcxx/src/expected.cpp
was added in https://github.com/llvm/llvm-project/pull/87390 and this
file assumes C++23 to be compiled. Apparently libc++ sources are always
built with C++23 so they don't guard things against it in `src/`:
https://github.com/llvm/llvm-project/pull/87390#discussion_r1575374945
This commit also builds libc++abi with C++23 because it doesn't seem
there's any downside to it.

- Exclude newly added `compiler_rt_shims.cpp`:
https://github.com/emscripten-core/emscripten/pull/22994/commits/5bbcbf0f5701583f3ac6a782f532a7ea9842cd21
We have excluded files in
https://github.com/emscripten-core/emscripten/tree/main/system/lib/libcxx/src/support/win32.
This is a new file added in this directory in
https://github.com/llvm/llvm-project/pull/83575.

- Disable time zone support:
https://github.com/emscripten-core/emscripten/pull/22994/commits/a5f2cbea31e398e63de004db521cb8e984dc8252
We disabled C++20 time zone support in LLVM 18 update (#21638):
https://github.com/emscripten-core/emscripten/commit/df9af64eb3f36b56133abacb6a0687253bd9563c
The list of source files related to time zone support has changed in
https://github.com/llvm/llvm-project/pull/74928, so this commit reflects
it.

- Re-add + update `__assertion_handler` from
`default_assertion_handler.in`:
https://github.com/emscripten-core/emscripten/pull/22994/commits/41f803725dc9d37063869a6a3653adb83fe0854c
This file was added as a part of LLVM 18 update (#21638) in
https://github.com/emscripten-core/emscripten/commit/8d51927d93fbdfb73b64298bd4bb1d3cee3cc3c8
and mistakenly deleted when I ran `update_libcxx.py`. This file was
copied from
https://github.com/llvm/llvm-project/blob/aadaa00de76ed0c4987b97450dd638f63a385bed/libcxx/vendor/llvm/default_assertion_handler.in,
so this also updates the file with the newest
`default_assertion_handler.in`.

- `_LIBCPP_PSTL_CPU_BACKEND_SERIAL` -> `_LIBCPP_PSTL_BACKEND_SERIAL`:
https://github.com/emscripten-core/emscripten/pull/22994/commits/4b969c3d3681410a79275ee4e93dc27fe58734c1
The name changed in this update, so reflecting it on our
`__config_site`.

- Directly include `pthread.h` from `emscripten/val.h`:
a5a76c3e3ae795a8a39cfe26bf7bc09aefbcbaf9
  Due to file rearrangements happened, this was necessary.

---

Other things to note:

- `std::basic_string<unsigned_char>` is not supported anymore
The support for `std::basic_string<unsigned_char>`, which was being used
by embind, is removed in this version.#23070 removes the uses from
embind.

- libcxxabi uses `__FILE__` in more places
https://github.com/llvm/llvm-project/pull/80689 started to use
`_LIBCXXABI_ASSERT`, which
[uses](https://github.com/llvm/llvm-project/blob/aadaa00de76ed0c4987b97450dd638f63a385bed/libcxxabi/src/abort_message.h#L22)
`__FILE__`, in `private_typeinfo.cpp`.
`__FILE__` macro produces different paths depending on the local
directory names and how the file is given to clang in the command line,
and this file is included in the result of one of our code size tests,
`other.test_minimal_runtime_code_size_hello_embind`, which caused the
result of the test to vary depending on the CI bots and how the library
is built (e.g., whether by embuilder, ninja, or neither).
Even though this was brought to surface during this LLVM 19 update,
`__FILE__` macro could be a problem for producing reproducible builds
anyway. We discussed this problem in #23195, and the fixes landed in
#23222, #23225, and #23256.