[link.py] Simplify process_libraries. NFC
These is no need for the second argument here since the only user of
this argument can simply use `new_flags` instead.  Both `new_flags` and
`linker_inputs` get combined in the link command in any case and they
are ordered by the index element of the tuples so that ordering doesn't
change here.