Reduce memory size used in test_memorygrowth_linear_step. NFC (#23301)
This doesn't change the test other than to use lower memory limits.

This change works around #22360 which seems to be some kind of bug in
node or v8. The result of the bug is that the node process becomes hung
in `uv_cond_wait` and never exists at the end of the test. Changing the
amount of memory used seems to work around the issue.