[compiler.mjs] Simplify file opening and finding logic. NFC
Removes the magic `find` function and instead require explicit file
locations at all the various call sites.  This means that in each
circumstance were we want to load a file there is only once place it
can/should be.

One side effect of this is that when using the `#include` preprocessor
directive the filename must either be absolute or relative to the
current file.