Remove legacy/broken makeCopyValues from parseTools_legacy.mjs.  NFC (#23354)
This function was calling functions that are not imported by this module
(at least isNumber and getFastValue). Rather than fix this by importing
them I think we can safely simple remove this.

Also, move isNumber in to parseTools.mjs which is its only user.

Also remove `stripCorrections` which was not exported via
`addToCompileTimeContext` is was not visible at all.