[EH] Add WASM_LEGACY_EXCEPTIONS option
This replaces the existing `WASM_EXNREF` option with
`WASM_LEGACY_EXCEPTIONS` option, in an effort to make the new standardized
exnref proposal the 'default' state and the legacy proposal needs to be
separately enabled an option. But given that most users haven't switched
to the new proposal and major web browsers haven't turned it on by
default, this `WASM_LEGACY_EXCEPTIONS` is turned on by default for the
moment.

This changes the following function names in `test/common.py` as well:
- `require(s)_wasm_eh` -> `require(s)_wasm_legacy_eh`
- `require(s)_wasm_exnref` -> `require(s)_wasm_eh`

The test names and suffixes will be changed in a follow-up.