FS: lookupPath should throw an error on an empty string
It's inconsistent that lookupPath returns `node:null` when the path
is an empty string. Normally it either raises `ENOENT` or `node` is truthy.
This makes empty string into an edge case that every caller has to handle.
`stat` has special handling for the empty string case and correctly returns
ENOENT on empty string, but chmod, truncate, and mknod all crash. utime and
some other functions mishandle the empty string in an unrelated way.