[EH] Add WASM_LEGACY_EXCEPTIONS option (#23365)
This replaces the existing `WASM_EXNREF` option with
`WASM_LEGACY_EXCEPTIONS` option, in an effort to make the new
standardized exnref proposal the 'default' state and the legacy proposal
needs to be separately enabled an option. But given that most users
haven't switched to the new proposal and major web browsers haven't
turned it on by default, this `WASM_LEGACY_EXCEPTIONS` is turned on by
default for the moment.

This changes the following function names in `test/common.py` as well:
- `require(s)_wasm_eh` -> `require(s)_wasm_legacy_eh`
- `require(s)_wasm_exnref` -> `require(s)_wasm_eh`

Even if `WASM_LEGACY_EXCEPTIONS` defaults to true for now, this adds
`set_setting(WASM_LEGACY_EXCEPTIONS)` in the tests to minimize future
fixes when `WASM_LEGACY_EXCEPTIONS` will eventually be turned off by
default.

The test names and suffixes will be changed accordingly in a follow-up.

This is in line with https://github.com/llvm/llvm-project/pull/122158.