[FS] Make fstatfs actually work
Prior to this change, `fstatfs` calls `statfs64` with the `char*` path
pointer argument set to zero. Generally byte zero of the memory seems
to be null, so when `statfs64` converts this `char*` to a JS string it
gets the empty string. Then it calls `FS.statfs` with the empty string
as an argument. `FS.statfs` calls `lookupPath` with an empty string
argument which returns a `null` node, so `fstatfs` returns the default
values for when there is no statfs handler in the files system.

This fixes the problem and adds testing for statfs/fstatfs on nodefs and
rawfs. We have to add separate `FS.statfsNode` and `FS.statfsStream` handlers
because in some cases `stream.node` is null and in some cases `stream.path
is useless (like for stdout). Also, in noderawfs we need to give paths for
the standard streams because there is unfortunately no `fs.fstatfsSync` so
we can only access information from the host via a path. This means the
noderawfs  test won't be portable to windows.