Remove `noProxy` option from `-sPROXY_TO_WORKER` (#23297)
The `noProxy` options was added back in 2014 in c2295c7b3, but doesn't
have any test coverage. This means there is is good chance its already
broken. What is more it a fair mount of complexity since it adds an
additional way in which certain JS files get pre-processed and included.

I propose that we simply remove it, but and alternative would be to fine
a way to test it. I'm not quite sure who the end user of this feature
might be though.