Only allow absolute paths for EM_CONFIG/EM_CACHE, etc.
Allowing relative paths here breaks when emcc is run as a subprocess
in a different directory.  This happens in when building system
libraries in batch build mode.

One alternative to doing this would be convert all these paths to
absolute paths when we read them and then re-export them with those
different values for our sub-processes, but that seems more complex and
error prone than just requiring them to be absolute in the first place.

I'm also not sure it makes much sense to use relative paths here since
if the environment contains a relative path then the compiler cannot be
used from different directories.  e.g.

```
$ export EM_CACHE=foo
$ emcc hello.c
$ cd foo && emcc foo.c  # confusingly would use a different cache
```

Fixes: #23374