Add `addAtPreRun` function for code injection at preRun.
This is a first step in trying to unify more of the code used by minimal
runtime and the normal runtime and also move all addOn* functions into
a library. Similar to the other addAt* functions for injecting code, this
adds `addAtPreRun`. `addAtPreRun` is used to have code that runs after the
Wasm is loaded, but before ctors.

The new injection point allows us to move `FS.init` into atPreRun and also
inline the call to `__wasm_call_ctors` at the same place in minimal runtime
and the normal runtime.