[EH] Rename EH library modes and file names (#23452)
- This renames `NoExceptLibrary` to `ExceptionLibrary`

- This renames the `Exception.WASM` to `Exceptions.WASM_LEGACY`, given
that we plan to add a new mode that uses the standard EH proposal soon,
which will be named `Exceptions.WASM`. This PR does not add the new mode
yet; it just renames what we already have.

- This makes `SjLjLibrary` use `eh_mode` as in `ExceptionLibrary`
instead of the current `is_wasm` boolean switch, given that we will need
to add another mode that uses the standard EH proposal to it soon. Its
`eh_mode` can't be set to `Exceptions.NONE` though.

- Library file suffixes changed
  - `ExceptionLibrary`
    - `NONE`: `-noexcept` -> `-noexcept` (no change)
    - `EMSCRIPTEN`: `` -> `` (no change)
    - `WASM_LEGACY`: `-except` -> `-legacyexcept`
    - `WASM` (will be added later): X -> `-wasmexcept`
  - `SjLjLibrary`
    - `EMSCRIPTEN`: `` -> `` (no change)
    - `WASM_LEGACY`: `-wasm-sjlj` -> `-legacysjlj`
    - `WASM` (will be added later): X -> `-wasmsjlj`

We give the current default mode (`EMSCRIPTEN`) no suffix. We hope we
eventually change `WASM` to be the default, in which case we will give
it no suffix and give `EMSCRIPTEN` a suffix like `-emexcept`/`-emsjlj`.

Also I removed / didn't add a `-` within a single suffix, so it is
`-legacyexcept` rather than `-legacy-except`. The reason is if a library
inherits from multiple classes, it will have multiple `-`s in its file
name like `libc++abi-debug-legacyexcept.a`, and I think it is clearer if
`-` only acts like split points between different suffixes.