[emcc.py] Simplify passing arguments to clang in compile-only mode (#23454)
This change avoids removing the `-o` flag and the input files when only
compiling. We delay the splitting out of the linker flags and the input
files until we know actually do to do linking.

This effect of this is that the compile-only phase is now much simply
because it doesn't need to re-inject the input files and the `-o` flag.

This also as the effect of keeping the order of those flags preserved
with respect to other command line flags when calling clang to do
compilation.