Remove `ATMAINS` from JS compiler
As far as I can tell the compile-time constant was never documented.

Unlike `ATEXITS` and `ATINITS` there was never any corresponding
`addAtMain`, and even the `ATMAIN` symbol itself was already not
available in `-sSTRICT` or `-sLEGACY_RUNTIME=0` mode.

It looks like it was specifically added in order to facilitate the
sealing of the filesystem in #8083, but its no longer used for that
purpose, and we have no current internal usages.