Unify more of the minimal runtime and normal runtime initialization code.
This is a first step in trying to unify more of the code used by minimal
runtime and the normal runtime and also move all addOn* functions into
a library.

First, make the call to the static constructors use the same code for both
modes. With this change, it becomes necessary to have an injection point
that runs before ctors and after ctors. To achieved that, I've replaced
the unused ATMAIN with new ATPOSTCTORS code injection point. The new
ATPOSTCTORS is similar to the old ATMAIN in that it runs after the static
constructors, but it now runs even if there is no main function.

Second, the FS initialization is now split into two parts (similar to how
it used to be before a refactoring).
1) atInit the std<X> stream are setup. This allows ctors to use printf etc.
2) atPostCtors the FS is locked down as late as possible but before main is
run.

Third, switch embind code generation to use the POSTCTORS so it can run
after the ctors have setup the bindings.

Fourth, minimal runtime changes slightly in that ATINIT runs before ctors.
This makes it consistent with the normal mode, but there is now POSTINIT
if anyone needs the old behavior.

Fifth, libdylink needs to use the new ATPOSTCTOR point to call the
static constructors in the secondary module so they are called after
the main module static constructors.