Fix createRequire so it works for local `node_modules`.
For most uses of `require()` in emscripten this does not matter since
we load mostly system libraries.  However for the `ws` module it is
needed.  This bug was being masked by the fact that we were setting
`NODE_PATH` in our socket test running.  This is no longer needed now
that we run (non-parallel) tests in the emscripten tree (in out/test).

This is followup to #23265 which itself was an attempt to revert 23169.

Fixes: #23503