Remove the need to jscompiler with CWD set to emscripten/src. NFC
The only place that CWD was relied upon here was the use of the `read()`
function in some JS library code that was expecting to be able to read
files relative to the `src` directory.  To deal with this use case I
created a new `readFile` function for compiler usage and modified the
exported `read` function to explicitly support `src/` relative files.

This change makes the existing `EMCC_BUILD_DIR` environment variable
obsolete since it was only needed because we were modifying CWD when
running the JS compiler.