Support closure in modularize instance mode.
ES module exports are added after closure is run, which leads to closure
complaining about undefined variables. To get around this, add the exports
to an extern file.

I've also changed MODULARIZE=instance to not require a `.mjs` output file.
This allows it to be used in test_codesize_minimal without changes and
some developers prefer not to use the `.mjs` extension, but use the setting
in package.json.