Update min python version from 3.6 to 3.8 (#23417)
The reason for picking 3.8 here is that it provides all the features we
currently have need of, and it available in the places we care about:

- debian/stable (bookworm): 3.11
- ubuntu/LTS (focal): 3.8
- emsdk: 3.9.2

The motivation for this change is twofold:

1. We get access to a few more python features.
2. We actually test using the version we claim to support (3.8 on
ubuntu/focal in CI). Currently the claim that we support 3.6 could well
be a lie. We don't do any testing on 3.6 do who knows if
incompatibilities have snuck in.

Replaces: #23378
Fixes: #23387