[EH] Make lsan/asan work with Wasm EH
`std::rethrow_exception` leaked memory because dependent exceptions were
not recognized as such and thus not freed properly here:
https://github.com/emscripten-core/emscripten/blob/44fd294e8244afa710cecb2c10c9a644a6549b8e/system/lib/libcxxabi/src/cxa_exception.cpp#L575-L583
The reason for that was, in `__cxa_rethrow_primary_exception`, we were
incorrectly throwing the primary exception, not the dependent one. I
think the error was introduced during copy-pasting.

`isDependentException` method was copied from `cxa_exception.cpp`,
because it is a static function, as in the case for other utility
functions in that file.

This also adds handling for dependent exceptions in
`__get_exception_message`.

Fixes #17004 and fixes #21124.