[test] Update GoogleTest tested against to version 1.15.2 (#23569)
The GoogleTest which is currently being tested works with emscripten as
part of the automated tests is a commit sometime after version 1.13.0
going by the readme when you cd into the submodule. This PR updates the
version being tested to the latest available release 1.15.2 . I updated
it to the tag v1.15.2 in GoogleTests tags, so it pinned to a specific
release.