Simplify linker flag handling. NFC (#23515)
This change simplifies the interface between emcc.py (the compiler
driver) and link.py (the linker).

Linker flags are no longer stored in the EmccState or stored alongside
their ordinal number. Instead passed directly the link.py as a simple
list.

This change brings us closer to de-coupling the linker process with the
hope of creating a standalone emld entry point.