Fix JSPI suspend through call from side module to main module (#23581)
This fixes a basic use case where a side module calls a suspending main
module symbol. Without this change, we call through a promising JS
wrapper produced by `instrumentWasmExports()` and fail to suspend.

We have to dlsym() the side module symbol since if we call it directly
it will fail because of another JS trampoline on the main -> side module
edge.