Move test_emscripten_log. NFC (#23613)
This test is the only thing left in the subdirectory. This change
also fixes search that was being done on the generated stackstrace.
The old code was simply searching for `_emscripten_log`, which happens
to be substring of `test_emscripten_log`. Since we are already using
more advanced regex below to extract the symbol name we can just use
that instead.