Detect and report late assignment of Module API elements
We had some bug reports recently when we switched to using async/await
in MODULUARIZE mode.  One of the side effects was the `--post-js` code
then runs after module instantiation and startup, which means assignment
of properties like `onRuntimeInitialized` is too late if it happens in
`--post-js`.

This was always the case for sync instantiation too.

We now detect this too-late-assignment at abort with a useful error.

See #23626 and #23420