Detect and report late assignment of Module API elements (#23629)
We had some bug reports recently when we switched to using async/await
in `MODULUARIZE` mode.  One of the side effects was the `--post-js` code
then runs after module instantiation and startup, which means assignment
of properties like `onRuntimeInitialized` is too late if it happens in
`--post-js`.

This was always the case for sync instantiation mode.

When assertions are enabled, we now detect this too-late-assignment and
abort
with a useful error.

Code size regression in this PR, as expected, only for debug builds.

See #23626 and #23420