Use stdin to pass settings to compiler.mjs / preprocessor.mjs      (#23650)
This avoids writing more files to disc than we need to.  One of my
hopes here is that this will be a little more performant, especially
on windows, where the filesystem isn't the fastest.

In order to aid debugging we stash copied of the these generated
settings using the save_intermediate mechanism.