[test] Add audio worklet parameter tests (and tidy other interactive tests) (#23659)
This is splits out the test changes from #23508 adding an interactive
parameter test:

```
test/runner interactive.test_audio_worklet_params_mixing
```

This test is also added to the browser tests now #23665 has merged,
offering a single test that touches all the various structs and their
offsets. When run as part of the `browser` tests it plays for 1 second,
for interactive without the `TEST_AND_EXIT` macro it will play
continuously.

It also tidies a little the shared code (and related tests) in the
process. Interactive tests had `_2gb` and `_4gb` options added (which
will fail until #23508 lands, but they're not part of the CI). All
features of these tests as interactive now work fully on Chrome, Firefox
and Safari (Safari has issues with looping, but it doesn't affect the
spirit of the test).