[AUDIO_WORKLET] Enable AW spinlocks, verify them in the browser tests (#23729)
- spinlocks fixed so they work in audio worklets
- adds tests for various spinlock cases
- test converted to C-only (still in a C++ file for now, see below)

This was unmerged last year (#22995), visited again now the browser
tests are running with audio. It's been moved over to `btest_exit()` and
verified that it runs (and will only exit after all tests have run or
asserted).

The only change I'd make is to rename the original test file, which has
grown beyond its original use. A `git mv` command won't survive multiple
commits, so I want to either leave it until last, live with the
resulting deletion, or rename as a separate PR.

Fixes: #22962