Avoid double-reporting the filename in JS library errors/warnings
As part of this I also avoid converting JS libraries path to absolute
paths where possible.  This means that errors and warnings show the
actual pathname passed by the users, and not some absolute path.  This
makes writing tests for the double-reporting issue easier, and makes the
errors/warnings easier to read (IMHO).