Remove unused JS library function: `jstoi_s`. NFC
This function was also mis-documented.  Its was added #10457 but even
then it was unused only add because "good to have for completeness".

However, the behavior of the JS `Number` function is not as described
in the comment.  Instead `Number("123abc")` returns NaN.

On balance I think its better to just remove it.