Add adjustWasmImports incoming Module setting
This provides a way to add symbols to the wasmImports dynamically.
This doesn't allow anything new compared to `instantiateWasm` except
that it doesn't interfere with sourcemaps or ASAN and it doesn't require
reproducing the somewhat tricky behavior in the standard `instantiateWasm`
function.

As an alternative, we could expose something like `instantiateWasmDefault`
to the user, and then they could do this like:
```js
Module['instantiateWasm'] = (imports, successCallback) => {
   adjustWasmImports(imports);
   instantiateWasmDefault(imports, successCallback);
}
```