[WasmFS] Create public `wasmfs_mount` function
We already had `wasmfs_unmount` so it seems logical we should have this
one too.

This is designed to replace `wasmfs_create_directory` which has a rather
misleading name.

I have left `wasmfs_create_directory` around for now but is simply
combination of `mkdir` + `mount` (which is essentially what is was
doing already).

This change ends up simplifying `doMkdir` which no longer needs to take
a backend argument.