[test] Remove test/wasmfs/get_backend.h. NFC
These days we have a better mechanism that uses  `setup_nodefs.js`
as a `--pre-js`.

See `setup_nodefs_test` in `test/common.py`

The advantage of this new mechanism is that it doesn't require the test
itself to know about wasmfs or any specific backend.

This test also adds `raw` to the list of wasmfs backends used by
`@wasmfs_all_backends` and `@also_with_wasmfs_all_backends`.  These is
a lot of overlap between these and other decorators in `common.py` but
we we can address that in a followup.