Add SDL3 CMake config file (#23869)
This is the equivalent of, and heavily based on, #15987 but for SDL3.

Assuming that I've understood
https://github.com/emscripten-core/emscripten/pull/19780#issuecomment-1622467552
correctly, none of the old style variables are expected to be set so we
only need to provide targets. This has the benefit that it Just Works
without needing to run `emcc -sUSE_SDL=3` first to download the port.