CI: Clean build directory on the library builder before persisting the workspace
The Ninja build of the libraries leaves behind all of the object files when
building the archives. This is nice for developers if you're working on the
libraries themselves but for CI, we are just bloating the persisted
workspace and not using the object files. So rather than automatically
clearing in embuilder or system_libs.py after the build, do it on CI.