Reduce non-highmem browser testing (#23911)
This reduces the number of tests run in the "default" browser test modes
(i.e. non-2g 32bit and non-4g 64-bit). The idea is that the set of edge
cases covered by the 2g and 4g modes is basically a superset of those
covered by the default cases. Tests that are disabled in the 2g and 4g
modes are still run in the default modes.