Use mainScriptUrlOrBlob if present under EXPORT_ES6 (#23890)
Spun out from #23804 .

This is a fix for #23769 which was introduced by #23171 , where the use
of the bundler pattern hard-coding a script name was made to apply to
all ES6 exports.

`mainScriptUrlOrBlob` is necessary under es6 exports to support running
a threaded emscripten module from another origin, e.g. to distribute an
emscripten module on a CDN (the script will load for the main thread but
won’t be allowed to create workers). Using blobs is an effective
workaround.