shell.js: Remove moduleOverrides. NFC (#23991)
This mechanism was added back in 80f517d78 so that shell code could
unconditionally assign to the `Module` object without clobbering the
incoming values there.

These days the shell code no longer assigns to the Module object so it
is no longer necessary.

Basically this code protects against accidental assignments to `Module`
between lines 138 and 408, but we don't have any such assignments that I
can see in our code.