[EH] Fix __get_exception_message for multiple inheritance (#24008)
#23552 added a routine to `__get_exception_message` so that in case of a
dependent exception (i.e., a rethrown exception using
`std::rethrow_exception`), it retrieves the correct primary exception's
pointer to call `what()` on. But this turns out not to be correct in all
cases, because `can_catch` can adjust the `thrown_object` pointer here:
https://github.com/emscripten-core/emscripten/blob/ea8eda50e93d28876907bd6576c705478f1c35da/system/lib/libcxxabi/src/cxa_exception_js_utils.cpp#L87-L94
Note that `adjustedPtr` parameter is a reference:

https://github.com/emscripten-core/emscripten/blob/ea8eda50e93d28876907bd6576c705478f1c35da/system/lib/libcxxabi/src/private_typeinfo.h#L25-L26

It appears one of the cases that `thrown_object` pointer can be adjusted
is when the caught exception is of a class that inherits from multiple
classes.

So in case of a dependent exception, if we just retrieve
`reinterpret_cast<__cxa_dependent_exception*>(exception_header)->primaryException`,
we lose the adjustment applied by `can_catch`.

This calls `can_catch` again to make the necessary adjustment again in
case of a dependent exception.

Fixes #23733.