Remove some Module property assertions. NFC
These are hindering work on other PRs such as #24005.  I've tried a
few other ways to unify and/or change these but simply removing this
mechanism and simplifying things seems like the cleanest approach.

All of the API for which were attaching this assertion have been removed
for well over a year so I think the loss is very minimal.