Improve performance of linker symbol cache under high contention. NFC
This drastically improves the performance of running the test suite
even though most users probably won't see any effect.

Prior to this change the symbol generation process itself always
happened with the symbol list lock held.  Now in the most common cases
the symbol list is generated before we grab the lock.

We really should look into a better way to do this without a global
symbol list lock, but this is a good interim step:

before:

- hot: 0m25.259s
- cold: 1m36.121s

after:

- hot: 0m20.220s
- cold: 0m22.324s