Allow <<< ATXXX >> code hooks in any part of the generated JS. NFC
This also simplify the writing of the final module to use the common
`write_file` utility.

I think in the past the generated JS could be so large that python could
have trouble processing it but I think that dates back the asm.js days
when the generated JS conatains all the asm.js generated code. These
days the size of the JS clue is never proportional to the size of the
input C/C++ program like it was back then.