Add support for passing a response file to `file_packager` (#24107)
Because calls to `file_packager` go through the shell wrapper, this can
on Windows due to a "command line is too long" error (e.g. with a long
list of `--preload-file`), even if a response file was provided to
`emcc` in order avoid the issue.

Fix it by supporting response files in `file_packager`, and changing
`link` so that it uses a response file where needed.