Respect runtime paths when loading shared libraries (#23872)
This updates `loadDynamicLibrary` to locate shared libraries from the file system
if possible.

When locating a shared library with a relative path, first attempt to find it in
the `LD_LIBRARY_PATH` directories. If it is not found there, search for it in the
runtime path directories. If it is not found, fall back to the original
`Module['locateFile']`.

Followup to #23805.

---------

Co-authored-by: Hood Chatham <roberthoodchatham@gmail.com>