Fix late-binding symbols with JSPI (implementation 2)
Late-binding symbols get a JS stub import that resolves the symbol
and then makes an onward call. This breaks JSPI.

This is a second approach to solving the problem by using WebAssembly.promising
and WebAssembly.Suspending with a JS trampoline. Unfortunately, as far as I can
tell there is no way to make it work for both a promising entrypoint and a
non-promising entrypoint. This is due to a change in the JSPI spec: early versions
said that if a suspending import does not return a promise, the suspender was allowed
to be null. However, the new version of JSPI eagerly traps if the suspender is null,
even if the function does not return a promise.