Disallow -sWASM_WORKERS with sanitizers (#24171)
Back in #21701 I moved the creation of wasmOffsetConvert and wasmOffset
into runtime_pthread.js but I didn't update libwasm_worker.js which
still uses the old name `wasmOffsetData` in its initial post message.

When trying to update this just now I noticed that the offset converter
(and santiizers in general) don't seem to work on `WASM_WORKERS` so this
change simply disables them.

The passing of `wasmOffsetData` and `wasmSourceMapData` was added back
in #20188 but it doesn't looks like support for santizers was supposed
to be part of that.